requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.30.0
pillow
motor
pydantic    